chore: improves API.
