+++ internal/collections/map_test.go
@@ -18,7 +18,6 @@ import (
 	"testing"
 
 	"github.com/corazawaf/coraza/v3/internal/regexp"
-
 	"github.com/corazawaf/coraza/v3/types/variables"
 )
