+++ internal/collections/named_test.go
@@ -8,7 +8,6 @@ import (
 	"testing"
 
 	"github.com/corazawaf/coraza/v3/internal/regexp"
-
 	"github.com/corazawaf/coraza/v3/types/variables"
 )
