+++ internal/operators/rx_test.go
@@ -7,10 +7,9 @@ import (
 	"fmt"
 	"testing"
 
-	"github.com/corazawaf/coraza/v3/internal/regexp"
-
 	"github.com/corazawaf/coraza/v3/experimental/plugins/plugintypes"
 	"github.com/corazawaf/coraza/v3/internal/corazawaf"
+	"github.com/corazawaf/coraza/v3/internal/regexp"
 )
 
 func TestRx(t *testing.T) {
